+++ tools/tfmanager/systembusdaemon.cpp
@@ -97,7 +97,7 @@ void SystemBusDaemon::readSocket()
     }
 
     QByteArray buf = socket->readAll();
-    tSystemDebug("SystemBusDaemon::read len : {}", (int64_t)buf.size());
+    tSystemDebug("SystemBusDaemon::read len : {}", (qint64)buf.size());
     if (maxServers <= 1) {
         // do nothing
         return;
@@ -161,7 +161,7 @@ void SystemBusDaemon::handleDisconnect()
         socketSet.remove(socket);
         disconnect(socket, nullptr, nullptr, nullptr);
         socket->deleteLater();
-        tSystemDebug("disconnected local socket : {:#x}", (uintptr_t)socket);
+        tSystemDebug("disconnected local socket : {:#x}", (quint64)socket);
     }
 }
